+++ code_snippets/content_cache.py
@@ -16,9 +16,13 @@ list_of_all_caches = ContentCache.list_caches(connection=conn)
 
 # Define a variable which can be later used in a script
 CACHE_ID = $cache_id  # Insert ID of cache on which you want to perform actions
+WH_TABLE_USED_ID = $wh_table_used_id  # Insert ID of WH table used by cache
 
 list_of_one_cache = ContentCache.list_caches(connection=conn, id=CACHE_ID)
 list_of_caches_as_dicts = ContentCache.list_caches(connection=conn, to_dictionary=True)
+list_of_caches_wh_table = ContentCache.list_caches(
+    connection=conn, wh_tables=WH_TABLE_ID
+)
 
 # Get single content cache by its id
 content_cache = ContentCache(connection=conn, cache_id=CACHE_ID)
@@ -35,6 +39,9 @@ content_cache.fetch()
 # Listing properties of content cache
 properties = content_cache.list_properties()
 
+# Invalidate content cache
+content_cache.invalidate()
+
 # Delete content cache
 content_cache.delete(force=True)
 
@@ -46,6 +53,11 @@ ContentCache.unload_caches(connection=conn, cache_ids=[CACHE_ID, OTHER_CACHE_ID]
 
 # Load multiple content caches
 ContentCache.load_caches(connection=conn, cache_ids=[CACHE_ID, OTHER_CACHE_ID])
+
+# Invalidate multiple content caches, we need to provide a list of cache ids
+# to be invalidated. Listing function returns a list of ContentCache objects
+to_be_invalid = [elem.id for elem in list_of_caches_wh_table]
+ContentCache.invalidate_caches(connection=conn, cache_ids=to_be_invalid)
 
 # Delete multiple content caches
 ContentCache.delete_caches(connection=conn, cache_ids=[CACHE_ID, OTHER_CACHE_ID])